Resolve "Buscar um usuário"

Closes #3 and #2

See merge request yuri-1684287759/notification-service!3

+++ src/main/java/br/com/vibbra/notificationservice/db/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends CrudRepository<UserEntity, Integer> {
+public interface UserRepository extends CrudRepository<UserEntity, Long> {
 
     boolean existsByEmail(final String email);
